Only import hcsshim package in windows files
Reduce packages that are compiled in linux binaries.
